requirements.txt
python-dotenv
pandas
numpy
certifi
scikit-learn
dill 
pymongo
pymongo[srv]
pyaml
mlflow
fastapi
uvicorn
dagshub
##-e .
